code clean up, etc
