+++ src/redux/reducers.js
@@ -6,6 +6,7 @@ import {
   REORDER_CARDS_IN_LIST,
   MOVE_CARD_TO_NEW_LIST
 } from './actions';
+import { dummyState } from './state';
 
 // eslint-disable-next-line
 Array.prototype.insert = function(index, item) {
@@ -21,47 +22,6 @@ export function handleBoard(state = {}, action) {
       return state;
   }
 }
-
-const dummyState = {
-  id: 'e97b8cd9-c2c0-455b-9324-ca4d0f5e959f',
-  title: 'Home',
-  lists: [
-    {
-      boardId: 'e97b8cd9-c2c0-455b-9324-ca4d0f5e959f',
-      listId: 'a820cd78-eee5-47ea-97cb-bfbcbc3e5750',
-      title: 'To Do',
-      cards: [
-        {
-          id: '4a4a6a4e-9a54-4234-b88f-cc998c375941',
-          listId: 'a820cd78-eee5-47ea-97cb-bfbcbc3e5750',
-          title: 'To Do Card 1'
-        },
-        {
-          id: '09ec1bd5-4994-4248-b7d5-8070bac8d5fe',
-          listId: 'a820cd78-eee5-47ea-97cb-bfbcbc3e5750',
-          title: 'To Do Card 2'
-        }
-      ]
-    },
-    {
-      boardId: 'e97b8cd9-c2c0-455b-9324-ca4d0f5e959f',
-      listId: '8560a8f5-e872-4ae5-b356-2d935af1b6a4',
-      title: 'Doing',
-      cards: [
-        {
-          id: '71c48add-4132-413b-954c-508f18a75e07',
-          listId: '8560a8f5-e872-4ae5-b356-2d935af1b6a4',
-          title: 'Doing Card 1'
-        },
-        {
-          id: 'cd39bd33-9afc-45df-ad28-6e1d01758c49',
-          listId: '8560a8f5-e872-4ae5-b356-2d935af1b6a4',
-          title: 'Doing Card 2'
-        }
-      ]
-    }
-  ]
-};
 
 export function boards(state = [dummyState], action) {
   switch (action.type) {
@@ -115,7 +75,7 @@ export function boards(state = [dummyState], action) {
                 ...list,
                 cards: list.cards.insert(action.destination, action.target)
               };
-            };
+            }
             return {
               ...list,
               cards: list.cards.splice(action.sourceId.index, 1)
