+++ src/redux/state.js
@@ -1,0 +1,40 @@
+export const dummyState = {
+  id: 'e97b8cd9-c2c0-455b-9324-ca4d0f5e959f',
+  title: 'Home',
+  lists: [
+    {
+      boardId: 'e97b8cd9-c2c0-455b-9324-ca4d0f5e959f',
+      listId: 'a820cd78-eee5-47ea-97cb-bfbcbc3e5750',
+      title: 'To Do',
+      cards: [
+        {
+          id: '4a4a6a4e-9a54-4234-b88f-cc998c375941',
+          listId: 'a820cd78-eee5-47ea-97cb-bfbcbc3e5750',
+          title: 'To Do Card 1'
+        },
+        {
+          id: '09ec1bd5-4994-4248-b7d5-8070bac8d5fe',
+          listId: 'a820cd78-eee5-47ea-97cb-bfbcbc3e5750',
+          title: 'To Do Card 2'
+        }
+      ]
+    },
+    {
+      boardId: 'e97b8cd9-c2c0-455b-9324-ca4d0f5e959f',
+      listId: '8560a8f5-e872-4ae5-b356-2d935af1b6a4',
+      title: 'Doing',
+      cards: [
+        {
+          id: '71c48add-4132-413b-954c-508f18a75e07',
+          listId: '8560a8f5-e872-4ae5-b356-2d935af1b6a4',
+          title: 'Doing Card 1'
+        },
+        {
+          id: 'cd39bd33-9afc-45df-ad28-6e1d01758c49',
+          listId: '8560a8f5-e872-4ae5-b356-2d935af1b6a4',
+          title: 'Doing Card 2'
+        }
+      ]
+    }
+  ]
+};
